mejoras
mejoras a la version anterior

+++ src/main/java/com/michaelpage/controller/EvaluacionRestController.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.michaelpage.models.entity.Evaluacion;
 import com.michaelpage.models.services.IEvaluacionService;
 
@@ -103,6 +102,12 @@ public class EvaluacionRestController {
 
 				evaluacionNew = evaluacionService.save(evaluacion);
 
+				if (evaluacionNew == null) {
+
+					throw new DataIntegrityViolationException("Faltan parametros");
+
+				}
+
 			}
 
 		} catch (Exception e) {
@@ -132,9 +137,14 @@ public class EvaluacionRestController {
 		try {
 			evaluacionActual = evaluacionService.findById(id);
 
+			if (evaluacionActual == null) {
+
+				throw new NoSuchElementException();
+			}
+
 		} catch (Exception e) {
 			// TODO: handle exception
-			response.put("mensaje", "no se encontró evaluación con el id ".concat(id.toString()).concat(" en la BD"));
+			response.put("mensaje", "no se encontro evaluación con el id ".concat(id.toString()).concat(" en la BD"));
 
 			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 
@@ -147,6 +157,12 @@ public class EvaluacionRestController {
 			evaluacionActual.setCalificacion(evaluacion.getCalificacion());
 			clienteUpdated = evaluacionService.save(evaluacionActual);
 
+			if (clienteUpdated == null) {
+
+				throw new DataIntegrityViolationException("faltan parametros");
+
+			}
+
 			response.put("mensaje", "los valores para la evaluacion ".concat(evaluacionActual.getNombre()
 					.concat(" y con el ID ").concat(id.toString().concat(" fueron editado con exito"))));
 			response.put("respuesta", clienteUpdated);
@@ -157,25 +173,40 @@ public class EvaluacionRestController {
 			// TODO: handle exception
 			response.put("mensaje", "no se pudo actualizar el registro indicado");
 			response.put("error", e.getMessage());
-			response.put("detalleError", e.getCause().toString());
-
 			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 
 		}
 
-	} 
+	}
 
 	@DeleteMapping("/eliminar/{id}")
 	public ResponseEntity<?> eliminarEvaluacionPorID(@PathVariable Long id) {
 
 		Map<String, Object> response = new HashMap<>();
 
-		try {
-			evaluacionService.delete(id);
-			response.put("mensaje",
-					"la evaluacion con el id ".concat(id.toString().concat(" se pudo eliminar con exito")));
+		Evaluacion evaluacionActual = new Evaluacion();
 
-			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.ACCEPTED);
+		try {
+
+			evaluacionActual = evaluacionService.findById(id);
+
+			if (evaluacionActual == null) {
+
+				throw new NoSuchElementException();
+			}
+
+			else
+
+			{
+
+				evaluacionService.delete(id);
+
+				response.put("mensaje",
+						"la evaluacion con el id ".concat(id.toString().concat(" se pudo eliminar con exito")));
+
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.ACCEPTED);
+
+			}
 
 		} catch (Exception e) {
